[protocol] Put all the Cerberus-specific files in their own directory
